Restore proper source code view in Storybook
