--- storybook/src/pages/internal/TablePage/stories/index.ts
@@ -1,2 +1,0 @@
-export { SortingWithSelect } from './SortingWithSelect'
-export { WithPagination } from './WithPagination'
